Extract station flattening code as a new method of CTSService
